rewriting certs Signed-off-by: Nikolay Demchuk <nikolay.demchuk@flant.com>
