+++ hooks/go/070-generate-certs/scheduler_extender_cert.go
@@ -16,8 +16,25 @@ limitations under the License.
 
 package generatecerts
 
-import "github.com/deckhouse/csi-nfs/hooks/go/certs"
+import (
+	"fmt"
+	"github.com/deckhouse/csi-nfs/hooks/go/consts"
+	tlscertificate "github.com/deckhouse/module-sdk/common-hooks/tls-certificate"
+)
 
 func init() {
-	certs.RegisterSchedulerExtenderCertHook()
+	var _ = tlscertificate.RegisterInternalTLSHookEM(tlscertificate.GenSelfSignedTLSHookConf{
+		CommonCACanonicalName: fmt.Sprintf("%s-%s", consts.ModulePluralName, consts.SchedulerCertCn),
+		CN:                    consts.SchedulerCertCn,
+		TLSSecretName:         fmt.Sprintf("%s-https-certs", consts.SchedulerCertCn),
+		Namespace:             consts.ModuleNamespace,
+		SANs: tlscertificate.DefaultSANs([]string{
+			consts.SchedulerCertCn,
+			fmt.Sprintf("%s.%s", consts.SchedulerCertCn, consts.ModuleNamespace),
+			fmt.Sprintf("%s.%s.svc", consts.SchedulerCertCn, consts.ModuleNamespace),
+			// %CLUSTER_DOMAIN%:// is a special value to generate SAN like 'svc_name.svc_namespace.svc.cluster.local'
+			fmt.Sprintf("%%CLUSTER_DOMAIN%%://%s.%s.svc", consts.SchedulerCertCn, consts.ModuleNamespace),
+		}),
+		FullValuesPathPrefix: fmt.Sprintf("%s.internal.customWebhookCert", consts.ModuleName),
+	})
 }
